dev-requirements.txt
python-dotenv==1.0.1
pydantic==2.7.0
odata1c-client==0.1.1
